Change all id fields to Long.

+++ PetAdoption-Backend/src/main/java/ca/mcgill/ecse321/petadoption/dao/AdvertisementRepository.java
@@ -3,8 +3,8 @@ package ca.mcgill.ecse321.petadoption.dao;
 import ca.mcgill.ecse321.petadoption.model.Advertisement;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AdvertisementRepository extends CrudRepository<Advertisement, Integer> {
+public interface AdvertisementRepository extends CrudRepository<Advertisement, Long> {
 
-    Advertisement findAdvertisementById(Integer id);
+    Advertisement findAdvertisementById(Long id);
 
 }
