+++ PetAdoption-Backend/src/main/java/ca/mcgill/ecse321/petadoption/dao/DonationRepository.java
@@ -3,8 +3,8 @@ package ca.mcgill.ecse321.petadoption.dao;
 import ca.mcgill.ecse321.petadoption.model.Donation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DonationRepository extends CrudRepository<Donation, Integer>{
+public interface DonationRepository extends CrudRepository<Donation, Long>{
 
-    Donation findDonationByTransactionNumber(Integer transactionNumber);
+    Donation findDonationByTransactionNumber(Long transactionNumber);
 
 }
